Core: fix CX86RecompilerOps::CompileLoadMemoryValue Map_GPR_32bit when called from LWC1

+++ Source/Project64-core/N64System/Mips/R4300iInstruction.cpp
@@ -199,7 +199,7 @@ void R4300iInstruction::ReadsGPR(uint32_t & Reg1, uint32_t & Reg2) const
     Reg2 = 0;
 }
 
-uint32_t R4300iInstruction::WritesGPR(void) const
+int32_t R4300iInstruction::WritesGPR(void) const
 {
     uint32_t op = m_Instruction.op;
     if (op == R4300i_SPECIAL)
@@ -226,7 +226,7 @@ uint32_t R4300iInstruction::WritesGPR(void) const
     {
         return 31; // RA
     }
-    return (uint32_t)-1;
+    return -1;
 }
 
 bool R4300iInstruction::ReadsHI() const
